Move text error component
refs #67
