+++ components/form-input-map/index.tsx
@@ -13,7 +13,7 @@ import OSM from 'ol/source/OSM';
 import VectorSource from 'ol/source/Vector';
 import { Circle as CircleStyle, Fill, Stroke, Style } from 'ol/style';
 import { useEffect, useRef, useState } from "react";
-import TextError from '../text-error';
+import TextError from '../text/error';
 
 // FUTURE: Add custom icon, optimize geolocation, and pan animation when map point is selected
 // https://openlayers.org/en/latest/examples/animation.html
